requirements.txt
boto3>=1.4.3
botocore>=1.4.93
PyYaml
hvac
mysql-connector-python
pyminifier
funcparserlib
iso8601
